Time: 161 ms (71.43%), Space: 107.1 MB (28.57%) - LeetHub

+++ 3075-maximize-happiness-of-selected-children/3075-maximize-happiness-of-selected-children.cpp
@@ -1,12 +1,11 @@
 class Solution {
 public:
     long long maximumHappinessSum(vector<int>& happiness, int k) {
-        // priority_queue<int,vector<int>,greater<int>> pq;
         sort(happiness.begin(),happiness.end());
         int decrement=0;
         long long ans=0;
         while(happiness.size() && k>0 && happiness.back()>0){
-            long long t=happiness.back()-decrement;
+            int t=happiness.back()-decrement;
             if(t<=0)
                 break;
             ans+=t;
